requirements.txt
Flask~=3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
Flask-Bootstrap==3.3.7.1
plotly==5.12.0
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
pytest~=8.3.4
Werkzeug~=2.3.8
jwt~=1.3.1
Flask-Login~=0.6.3
SQLAlchemy~=2.0.37
WTForms~=3.2.1